backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
torch>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
huggingface_hub>=0.20.0
scipy>=1.10.0
